test(composio): move tests to *_test files
