+++ src/openhuman/composio/ops.rs
@@ -1867,7 +1867,7 @@ pub async fn composio_clear_api_key(config: &Config) -> OpResult<RpcOutcome<serd
 }
 
 #[cfg(test)]
-#[path = "ops_tests.rs"]
+#[path = "ops_test.rs"]
 mod tests;
 
 // ── Helpers re-exported so callers can pull connection/tool types without
